requirements.txt
aiodns==3.2.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
attrs==23.2.0
beautifulsoup4==4.12.3
cachetools==5.3.3
ccxt==4.3.27
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
distro==1.9.0
duckduckgo_search==6.1.0
frozendict==2.4.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
googleapis-common-protos==1.63.0
greenlet==3.0.3
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
lxml==5.2.2
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.0.5
multitasking==0.0.11
numpy==1.26.4
oauthlib==3.2.2
ollama==0.2.0
openai==1.30.1
orjson==3.10.3
pandas==2.2.2
peewee==3.17.5
pgvector==0.2.5
phidata==2.4.8
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycares==4.4.0
pycparser==2.22
pydantic==2.7.1
pydantic-settings==2.2.1
pydantic_core==2.18.2
Pygments==2.18.0
pyparsing==3.1.2
pyreqwest_impersonate==0.4.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
setuptools==69.5.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
tavily-python==0.3.3
tiktoken==0.7.0
tomli==2.0.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
webencodings==0.5.1
wikipedia==1.4.0
yarl==1.9.4
yfinance==0.2.38
youtube-transcript-api==0.6.2
